submissions/76-cerebras-inference/requirements.txt
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
